interfaces/xmpp-client/requirements.txt
dnspython
xmpppy
